src/requirements.txt
astroid==2.3.3
Babel==2.7.0
Click==7.0
Django==2.2.10
django-cors-headers==3.2.0
django-dotenv==1.4.2
django-phonenumber-field==3.0.1
djangorestframework==3.10.3
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pdfkit==0.6.1
phonenumbers==8.11.0
Pillow==6.2.2
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
gunicorn
whitenoise
dj-database-url
django-rest-swagger==2.1.2
coreapi>=2.3.3
coreschema>=0.0.4
ruamel.yaml>=0.15.34
flake8
pytest
inflection>=0.3.1
uritemplate>=3.0.0
packaging
drf-yasg==1.17.1
git+git://github.com/johnfraney/wkhtmltopdf-pack.git
django-rest-passwordreset==1.1.0
sendgrid>=6.2.0
django-filter>=2.2.0
django-notifications-hq